# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IT-Ticket API
  version: 1.1.0
  description: >
    REST-API för IT-Ticket — ett ärendehanteringssystem för intern IT-support på Prefabmästarna.


    **Leverans:** rullande från `main` (en instans per deployment).
    Versionen ovan spårar API-kontraktet, inte appens git-tagg.


    ## API coverage

    This spec documents the full API surface: 160 operations
    across every route file — health/CSRF, auth, tickets, categories,
    contacts, companies, kb, public, users, API keys,
    settings, comments, attachments, shares, links, checklists, checklist
    templates, template fields, tags,
    reports, backup, push notifications, webhooks, and inbound email.


    **Prose reference:** `docs/API.md` — generated from the actual Express
    routers in `server/src/routes/` and kept up to date on route changes. Use
    it as a narrative cross-check alongside this machine-readable spec.


    ## Autentisering

    Två sätt att autentisera, båda via `Authorization`-headern:

    - **Session (JWT):** `Authorization: Bearer <jwt>`. Access-token gäller 15 min;
      en HttpOnly refresh-cookie (path `/api/auth`) förnyar den via
      `POST /api/auth/refresh`. **Muterande anrop (POST/PUT/PATCH/DELETE) kräver
      även CSRF-token** i `x-csrf-token` (hämtas från `GET /api/csrf-token`).
    - **API-nyckel:** `Authorization: Bearer itk_live_...`. Nycklar har scope
      `read`, `read,write` eller — för administrativa endpoints — även `admin`.
...


...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /webhooks:
    get:
      operationId: webhooksList
      tags: [Webhooks]
      summary: List configured outgoing webhooks
      description: Admin only. The `secret` (HMAC signing key) is never returned.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Webhooks (without secret)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WebhooksWebhook" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: webhooksCreate
      tags: [Webhooks]
      summary: Create an outgoing webhook
      description: >
        Admin only. `url` is validated against SSRF (see
        `lib/webhookValidator.ts`; blocks e.g. internal/loopback targets) and
        re-validated at delivery time to defend against DNS-rebinding.
        `events` must be a non-empty subset of the fixed event list. A
        32-byte random hex `secret` is generated and stored in **readable
        (not hashed) form** — by design, since it is an HMAC signing key that
        both this server and the receiving endpoint need in raw form to
        compute/verify signatures (the same pattern used by GitHub and
        Stripe), not a password. The `secret` is returned **only in this
        create response** — subsequent `GET`/`PUT` responses omit it.


        **Delivery &amp; signature verification** (not itself an HTTP
        endpoint on this API — this is what the *receiving* server, at
        `url`, must implement). On each matching event this server POSTs a JSON
        body `{ "event": "...", "payload": { ... }, "timestamp": "<ISO>" }` with
        these headers:

        - `X-Webhook-Id` - delivery UUID, **stable across retries** (dedupe on it).
        - `X-Webhook-Timestamp` - unix seconds, set **per attempt**.
        - `X-Webhook-Event` - the event type string.
        - `X-Webhook-Signature` - `hex(HMAC-SHA256(secret, timestamp + "." + id + "." + rawBody))`.

        To verify: recompute the HMAC over the **raw request bytes** (not a
        re-serialised JSON object), compare in constant time, reject timestamps
        older than ~5 minutes, and ignore ids you have already processed.
        **Breaking change:** the previous scheme signed only the body; that
        signature no longer verifies.


        Redirects are never followed - a 3xx response is a final failure (no
        retry). Other non-2xx responses and network/DNS-safety failures retry up
        to 5 attempts in total with back-off of 1m, 5m, 30m and 2h between attempts, then stop.
        `ticket.updated` is only emitted when the status changes, with the
        minimal payload `{ id, status, priority, assigned_to, title,
        updated_fields }` (notes and solution text never leave the system).
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    enum:
                      - ticket.created
                      - ticket.updated
                      - ticket.closed
                      - ticket.deleted
                      - ticket.status_changed
                      - comment.created
                      - contact.created
                      - contact.updated
      responses:
        "201":
          description: Created webhook, including the one-time-visible secret
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhooksWebhookWithSecret" }
        "400":
          description: Missing/unsafe `url`, or missing/invalid `events`
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Error" }
  /webhooks/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    put:
      operationId: webhooksUpdate
      tags: [Webhooks]
      summary: Update a webhook (partial)
      description: >
        Admin only. Any of `url`, `events`, `active` may be omitted to leave
        unchanged; a new `url` is SSRF-validated the same way as creation.
        The secret cannot be changed or re-viewed through this endpoint.
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
                active: { type: boolean }
      responses:
        "200":
          description: Updated webhook (without secret)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhooksWebhook" }
        "400":
          description: Unsafe `url`
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Error" }
    delete:
      operationId: webhooksDelete
      tags: [Webhooks]
      summary: Delete a webhook
      description: Admin only.
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Session-JWT: `Authorization: Bearer <jwt>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API-nyckel: `Authorization: Bearer itk_live_...` (scope read/write/admin; admin-skyddade endpoints kräver `admin` även när ägaren är administratör). Undantagen från CSRF."
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Krävs på muterande sessions-anrop. Hämtas från `GET /api/csrf-token`.
  responses:
    Error:
      description: Fel
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Message:
      description: OK med meddelande
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
    Unauthorized:
      description: Ej autentiserad
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Saknar behörighet (t.ex. admin krävs eller fel scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Hittades inte
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    WebhooksWebhook:
      type: object
      description: Webhook row without the HMAC secret.
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        events: { type: string, description: JSON-encoded array of event-type strings. }
        active: { type: integer, enum: [0, 1] }
        created_at: { type: string, format: date-time }
        last_triggered_at: { type: string, format: date-time, nullable: true }
    WebhooksWebhookWithSecret:
      allOf:
        - $ref: "#/components/schemas/WebhooksWebhook"
        - type: object
          properties:
            secret:
              type: string
              description: >
                32-byte hex HMAC-SHA256 signing key. Returned only on
                creation — store it now, it cannot be retrieved again.
...